Convert the dnd-draggable link function into a directive controller
To be able to have a `dnd-handle` directive that designates one or more
elements inside a `dnd-draggable`, the latter needs to expose an API
and thus a directive controller.

This is a preparatory commit with no functional change that only
converts the link function into a controller definition, entailing lots
of added dollar signs and indentation changes.