Implementation of dnd-handle directive
This commit adds a new `dnd-handle` directive, and the code to support
these handles in the existing `dnd-draggable` directive.

The basic design is this:
 * Any element decorated with the `dnd-handle` directive registers
itself with the parent `dnd-draggable` directive controller.
 * Also, it listens for `mousedown` and `mouseup` events in the handle
element. These event listeners control the `handleGrabbed` flag in the
`dnd-draggable` directive.
 * When the actual `dragstart` event is fired on the draggable element,
the `dnd-draggable` directive checks whether it has any registered
handles, and if so, only starts the drag if any one them have been
grabbed (i.e. the `handleGrabbed` flag is set).