Insert item at correct index for filtered lists.
With this change, the moved/copied item is not placed with the index of the placerholder in the DOM list, but rather by comparing the indices of the elements next to the placeholder. This way, if you have a list where you filter out elements, you still insert the dropped elements at the correct position.