Find the best position of place holder where event target is the list itself
Suppose all children of a dnd-list have margin and you are about to insert a new item between them. In this case, the event target and the list are the same and the placeholder will be put at the end of the list.

In this commit, a correct location for both placeholder and the new element is computed based on the dnd-list itself.